+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Moved from using the Jaeger exporter to OTLP exporter by default
 * Moved to Java 11 at language level, dropped Java 8 and use Java 17 as the runtime for all containers
+* Fixed bug about missing OAuth password grants configuration   
 * Dependency updates (Vert.x 4.3.5)
 * Dependency updates (snakeYAML [CVE-2022-41854](https://nvd.nist.gov/vuln/detail/CVE-2022-41854))